Move config definition to subnamespace
